Adds ORM to DAL and db access to items endpoint
 ## Description
Using KTORM adds ORM to the data layer. The implementation is based on a Postgres database whose connections are managed by hikariCP. As for the database schema management, flyway was added to the project to apply schema updates through migration scripts. The chosen approach for the ORM architecture was to map table objects directly to domain entity classes. The previously implemented endpoint "List Hotelier Items" is now functioning with database access.

 ### Changes Overview
- `docker-compose.yml` updated with the description of the postgres container
- `application.hotelier` package created containing the `HotelierDTO` and `IHotelierService` classes
- Domain entities updated to fully reflect the database schema
- Creates enums for item category and reputation badge
- Creates the `dal.db.DbFactory` class to instantiate the ktorm database object using hikariCP postgres datasource and ensure flyway migrations applying.
- Creates the table objects for Hoteliers and Items